code/requirements.txt
zipFile
requests
pytorch>=1.10.0
torchvision>=0.11.0
